Lint, Reorganize, Rename (Batch commit)
