--- client/src/components-style/ReviewsBox-style.jsx
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-export const ReviewsLength = styled.div`
-  position: relative;
-  font-size: 17px;
-  font-weight: 600;
-`;
