+++ client/src/components/App/AppStyles.jsx
@@ -1,0 +1,23 @@
+import styled from 'styled-components';
+
+export const AppBody = styled.div`
+
+`;
+
+export const Title = styled.h3`
+  padding-left: 50px;
+  position: relative;
+`;
+
+
+export const AllItems = styled.div`
+  margin-top: -36px;
+  padding-bottom: 15px;
+  left: 900px;
+  position: relative;
+  font-size: 15px;
+  color: #0073bb;
+  &:hover {
+    text-decoration: underline;
+  }
+`;
