+++ client/src/components/Review/ReviewStyles.jsx
@@ -1,0 +1,73 @@
+import styled from 'styled-components';
+
+export const SingleReview = styled.div`
+  position: relative;
+`;
+
+export const PhotoCropper = styled.div`
+  margin-top: 15px;
+  position: relative;
+  width: 30px;
+  height: 30px;
+  border-radius: 3px;
+  overflow: hidden;
+`;
+
+export const ProfilePic = styled.img`
+  height: 30px;
+`;
+
+export const UserName = styled.div`
+  position: relative;
+  margin-top: -32.5px;
+  margin-left: 35px;
+  font-weight: 600;
+  font-size: 14px;
+`;
+
+export const TinyPairIcon = styled.div`
+  position: relative;
+  margin-left: 35px;
+  fill: orangered;
+`;
+
+export const FriendCount = styled.div`
+  position: relative;
+  margin-left: 58px;
+  font-size: 12px;
+  color: gray;
+  font-weight: 600;
+  margin-top: -19.05px;
+`;
+
+export const TinyStar = styled.div`
+  position: relative;
+  margin-top: -14px;
+  margin-left: 80px;
+`;
+
+export const UserReviews = styled.div`
+  position: relative;
+  margin-top: -14px;
+  margin-left: 100px;
+  font-size: 12px;
+  color: gray;
+  font-weight: 600;
+`;
+
+export const StarContainer = styled.div`
+  position: relative;
+  margin-top: 10px;
+`;
+
+export const Date = styled.div`
+  position: relative;
+  margin-left: 95px;
+  font-size: 14px;
+  color: gray;
+  font-weight: 400;
+`;
+export const ReviewText = styled.div`
+  position: relative;
+  margin-top: 10px;
+`;
