+++ client/src/components/ReviewsBox/ReviewsBoxStyle.jsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const ReviewsLength = styled.div`
+  position: relative;
+  font-size: 17px;
+  font-weight: 600;
+`;
