--- client/src/components/camera.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Camera = () => {
-  return (
-    <div>
-      <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"> <path d="M 9.28125 3 C 8.179688 3 7.230469 3.808594 7.007813 4.929688 L 6.8125 6 L 4.351563 6 C 3.050781 6 2 7.050781 2 8.339844 L 2 18.671875 C 2 19.960938 3.050781 21 4.339844 21 L 19.660156 21 C 20.949219 21 22 19.960938 22 18.671875 L 22 8.328125 C 22 7.039063 20.949219 6 19.660156 6 L 17.1875 6 L 16.988281 4.929688 C 16.769531 3.808594 15.820313 3 14.71875 3 Z M 11.96875 9 C 14.511719 9 16.570313 11.011719 16.570313 13.5 C 16.570313 15.988281 14.511719 18 11.96875 18 C 9.429688 18 7.367188 15.988281 7.367188 13.5 C 7.367188 11.011719 9.429688 9 11.96875 9 Z M 11.96875 11 C 10.53125 11 9.367188 12.121094 9.367188 13.5 C 9.367188 14.878906 10.53125 16 11.96875 16 C 13.402344 16 14.570313 14.878906 14.570313 13.5 C 14.570313 12.121094 13.402344 11 11.96875 11 Z" /></svg>        
-    </div>
-  )
-}
-
-export default Camera;
